created interpreter struct (will be needed for next section) + removed warnings
